+++ src/main/java/Trendithon/SpinOff/domain/dto/SignUpDto.java
@@ -1,4 +1,4 @@
-package Trendithon.SpinOff.web.dto;
+package Trendithon.SpinOff.domain.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
